Piece Functionality
- Piece uses arraylist of Integer to keep track of possible moves of
pieces. Can convert back from integer to coordinate using: x = n%8 and y
= n/8. Additionally, keeps track if the piece has moved or not which is
needed for certain moves such as pawn double square advance or
casteling. Each subclass implements updateReach which is called after
the motion of any piece for all the pieces on the board. setReach is a
helper method which is used by Rook, Bishop and Queen.
- Pawn has regular moves + capture functionality however enpassant and
promotion still needs to be implemented.
- King has regular moves functional however casteling and
check/checkmate conditions still need to be implemented.
- Bishop, Queen, Rook and Knight are fully implemented in their moves.
- need to implement a way to enforce pins.